testfrominstance/chalicelib/auth_helper.py
import random
from chalice import Chalice, AuthResponse
from chalice import CORSConfig
import psycopg2
import json
import requests
import sendgrid
import os
from sendgrid.helpers.mail import *
from jose import jwt
from functools import wraps
from six.moves.urllib.request import urlopen
from app import get_token_auth_header
from chalicelib.aws_secrets_helper import get_secret


AUTH0_DOMAIN = json.loads(get_secret())["auth0Domain"]
API_AUDIENCE = json.loads(get_secret())["auth0AudienceUrl"]
ALGORITHMS = ["RS256"]




def requires_scope(required_scope, token):
    """Determines if the required scope is present in the Access Token
    Args:
        required_scope (str): The scope required to access the resource
    """

    unverified_claims = jwt.get_unverified_claims(token)
    if unverified_claims.get("scope"):
            token_scopes = unverified_claims["scope"].split()
            for token_scope in token_scopes:
                if token_scope == required_scope:
                    return True
    return False



def requires_auth_mentee(f):
    """Determines if the Access Token is valid
    """
    
    @wraps(f)
    def decorated(*args, **kwargs):
        token = get_token_auth_header()
        jsonurl = urlopen("https://"+AUTH0_DOMAIN+".well-known/jwks.json")
        jwks = json.loads(jsonurl.read())
        unverified_header = jwt.get_unverified_header(token)
        rsa_key = {}
        for key in jwks["keys"]:
            if key["kid"] == unverified_header["kid"]:
                rsa_key = {
                    "kty": key["kty"],
                    "kid": key["kid"],
                    "use": key["use"],
                    "n": key["n"],
                    "e": key["e"]
                }
        if rsa_key:
            try:
                payload = jwt.decode(
                    token,
                    rsa_key,
                    algorithms=ALGORITHMS,
                    audience=API_AUDIENCE,
                    issuer="https://"+AUTH0_DOMAIN
                )

            except Exception as e:
                print(e)
                raise Exception

            if(requires_scope("edit:mentee", token)):
                return f(*args, **kwargs)
    return decorated



def requires_auth_mentor(f):
    """Determines if the Access Token is valid
    """
    @wraps(f)
    def decorated(*args, **kwargs):
        token = get_token_auth_header()
        jsonurl = urlopen("https://"+AUTH0_DOMAIN+".well-known/jwks.json")
        jwks = json.loads(jsonurl.read())
        unverified_header = jwt.get_unverified_header(token)
        rsa_key = {}
        for key in jwks["keys"]:
            if key["kid"] == unverified_header["kid"]:
                rsa_key = {
                    "kty": key["kty"],
                    "kid": key["kid"],
                    "use": key["use"],
                    "n": key["n"],
                    "e": key["e"]
                }
        if rsa_key:
            try:
                payload = jwt.decode(
                    token,
                    rsa_key,
                    algorithms=ALGORITHMS,
                    audience=API_AUDIENCE,
                    issuer="https://"+AUTH0_DOMAIN
                )

            except Exception as e:
                print(e)
                raise Exception

            if(requires_scope("edit:mentor", token)):
                return f(*args, **kwargs)
    return decorated


def requires_auth(f):
    """Determines if the Access Token is valid
    """
    @wraps(f)
    def decorated(*args, **kwargs):
        token = get_token_auth_header()
        jsonurl = urlopen("https://"+AUTH0_DOMAIN+".well-known/jwks.json")
        jwks = json.loads(jsonurl.read())
        unverified_header = jwt.get_unverified_header(token)
        rsa_key = {}
        for key in jwks["keys"]:
            if key["kid"] == unverified_header["kid"]:
                rsa_key = {
                    "kty": key["kty"],
                    "kid": key["kid"],
                    "use": key["use"],
                    "n": key["n"],
                    "e": key["e"]
                }
        if rsa_key:
            try:
                payload = jwt.decode(
                    token,
                    rsa_key,
                    algorithms=ALGORITHMS,
                    audience=API_AUDIENCE,
                    issuer="https://"+AUTH0_DOMAIN
                )

            except Exception as e:
                print(e)
                raise Exception

            if(requires_scope("edit:mentor", token) or requires_scope("edit:mentee", token) ):
                return f(*args, **kwargs)
    return decorated
