testfrominstance/requirements.txt
attrs==17.4.0
botocore==1.12.75
certifi==2018.11.29
chalice==1.6.2
chardet==3.0.4
click==6.7
docutils==0.14
ecdsa==0.13
enum-compat==0.0.2
future==0.17.1
idna==2.8
jmespath==0.9.3
psycopg2==2.7.6.1
pyasn1==0.4.5
PyCronofy==1.1.7
PyJWT==1.7.1
PySocks==1.6.8
python-dateutil==2.7.5
python-http-client==3.1.0
python-jose==3.0.1
pytz==2018.9
requests==2.21.0
rsa==4.0
sendgrid==5.6.0
six==1.12.0
SQLAlchemy==1.2.15
twilio==6.22.1
typing==3.6.4
urllib3==1.24.1
boto3
